qt6: Handle size() change from int to size_t.
None of these values should ever be large enough to overflow an int,
so add a static_cast to make the compiler happy.

+++ mythtv/libs/libmythtv/HLS/httplivestreambuffer.cpp
@@ -830,7 +830,7 @@ class HLSStream
             // not even size, pad with front 0
             line.insert(2, QLatin1String("0"));
         }
-        int padding = max(0, AES_BLOCK_SIZE - (line.size() - 2));
+        int padding = max(0, AES_BLOCK_SIZE - (static_cast<int>(line.size()) - 2));
         QByteArray ba = QByteArray(padding, 0x0);
         ba.append(QByteArray::fromHex(QByteArray(line.toLatin1().constData() + 2)));
         std::copy(ba.cbegin(), ba.cend(), m_aesIv.begin());
